CI: test against go1.22 (#1195)
This PR modifies CI to test against Go versions 1.21 and 1.22. To avoid
forcing users to upgrade to 1.21, I kept `go.mod` version at `1.20.`

Some tests had to be updated to pass in 1.22.

I can remove `stack_120_test.go` if there's a good argument for it, but
my thoughts are that we should have a complete & passing test suite for
every version >= `go.mod`'s version.