fx.Self: a parameter to fx.As for providing a type as itself
We frequently see requests for folks who want to use `fx.As`
to provide a type as another type, while also providing it as itself:
* https://github.com/uber-go/fx/discussions/1196
* https://github.com/uber-go/fx/discussions/1148
* https://github.com/uber-go/fx/discussions/1079

This is currently not possible because `fx.As` causes a constructor
to no longer provide its original type.

The workaround we often give is for folks to do something like:
```go
fx.Provide(
    newConcreteType,
    func(ct *concreteType) Interface {
        return ct
    }
)
```
which is admitedly not very ergonomic, and leads to a common mistake
of providing the cosntructor twice instead:
```go
fx.Provide(
    newConcreteType,
    fx.Annotate(newConcreteType, fx.As(new(Interface))),
)
```

This PR adds `fx.Self()`, which returns a special value to indicate
an `fx.As` should retain that original return type:
```go
fx.Provide(
    newConcreteType,
    fx.As(fx.Self()),
    fx.As(new(Interface)),
)
```

This can be used as positional arguments to `fx.As` as well.