Fix deadlock caused by race while signal receivers are stopping
A user reported a possible deadlock within the signal receivers (#1219).

This happens by:
* `(*signalReceivers).Stop()` is called, by Shutdowner for instance.
* `(*signalReceivers).Stop()` acquires the lock.
* Separately, an OS signal is sent to the program.
* There is a chance that `relayer()` is still running at this point
  if `(*signalReceivers).Stop()` has not yet sent along the `shutdown` channel.
* The relayer attempts to broadcast the signal received via the `signals` channel.
* Broadcast()` blocks on trying to acquire the lock.
* `(*signalReceivers).Stop()` blocks on waiting for the `relayer()` to finish
  by blocking on the `finished` channel.
* Deadlock.

Luckily, this is not a hard deadlock, as `Stop` will return if the context times out.

This PR fixes this deadlock. The idea behind how it does it
is based on the observation that the broadcasting logic does not
necessarily seem to need to share a mutex with the rest of `signalReceivers`.
Specifically, it seems like we can separate protection around
the registered `wait` and `done` channels, `last`, and the rest of the fields.
To avoid overcomplicating `signalReceivers` with multiple locks for different uses,
this PR creates a separate `broadcaster` type in charge of keeping track of
and broadcasting to `Wait` and `Done` channels.

Having a separate broadcaster type seems actually quite natural,
so I opted for this to fix the deadlock. Absolutely open to feedback
or taking other routes if folks have thoughts.

Since broadcasting is protected separately, this deadlock no longer happens
since `relayer()` is free to finish its broadcast and then exit.

In addition to running the example provided in the original post to verify,
I added a test and ran it before/after this change.

Before:
```
$ go test -v -count=10 -run "TestSignal/stop_deadlock" .
=== RUN   TestSignal/stop_deadlock
    signal_test.go:141:
                Error Trace:
/home/user/go/src/github.com/uber-go/fx/signal_test.go:141
                Error:          Received unexpected error:
                                context deadline exceeded
                Test:           TestSignal/stop_deadlock
```
(the failure appeared roughly 1/3 of the time)

After:
```
$ go test -v -count=100 -run "TestSignal/stop_deadlock" .
--- PASS: TestSignal (0.00s)
    --- PASS: TestSignal/stop_deadlock (0.00s)
```
(no failures appeared)