Update modules.md
I don't think the example follows good coding conventions. If `Config` is supposed to be a parameter object, it should be named `Params` since the constructor is called `New(...)` and needs defined `fx.In` within. If it is not meant to be an parameter object, I think passing `Params` into the constructor without defining that in the example is confusing.