WIP: [PoC] fx.Evaluate
**This is just a PoC at this time and not yet ready for review.**

## Details

This change introduces the `fx.Evalute` option,
which allows for dynamic generation of provides, invokes, decorates,
and other evaluates based on the state of existing dependencies.

It addresses one of the more significant remaining pain points
with usage of Fx in large applications and ecosystems:
the ability to programmatically generate parts of the dependency graph.

Concretely, this makes the following possible:

```
fx.Evaluate(func(cfg *Config) fx.Option {
    if cfg.Environment == "production" {
        return fx.Provide(func(*sql.DB) Repository {
            return &sqlRepository{db: db}
        }),
    } else {
        return fx.Provide(func() Repository {
            return &memoryRepository{}
        })
    }
}),
fx.Provide(func(...) *sql.DB { ... }),
```

With fx.Evaluate, the dependency on `*sql.DB` is present in the graph
only in production environments.
In development environments, the dependency connection is absent,
and therefore the database connection is never established.

(Today, attempting to write an Fx module that switches on the backend
like this will result in the SQL connection being established even in
development environments.)

**TODO**

- [ ] Decide whether nilling out the slice is the right approach
      for tracking what's been already run
- [ ] fx.Module handling
- [ ] fx.WithLogger handling
- [ ] Testing with every other feature
- [ ] Documentation